# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateSprintRequest:
    properties:
      end_date:
        description: EndDate is the date when the sprint ends.
        example: "2025-04-30T00:00:00Z"
        type: string
      goal:
        description: Goal is the objective or goal of the sprint.
        example: Complete initial UI design
        minLength: 5
        type: string
      name:
        description: Name is the name of the sprint.
        example: Sprint 1
        minLength: 2
        type: string
      project_id:
        description: ProjectID is the ID of the project this sprint belongs to.
        example: 1
        minimum: 1
        type: integer
      start_date:
        description: StartDate is the date when the sprint begins.
        example: "2025-04-15T00:00:00Z"
        type: string
    required:
    - end_date
    - goal
    - name
    - project_id
    - start_date
    type: object
...
  dto.ErrorResponse:
    properties:
      details: {}
      message:
        example: An error occurred
        type: string
    type: object
  dto.GenericSuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  dto.SprintResponse:
    properties:
      end_date:
        description: EndDate is the date when the sprint ends.
        example: "2025-04-30T00:00:00Z"
        type: string
      goal:
        description: Goal is the objective or goal of the sprint.
        example: Complete initial UI design
        type: string
      id:
        description: ID is the unique identifier of the sprint.
        example: 1
        type: integer
      name:
        description: Name is the name of the sprint.
        example: Sprint 1
        type: string
      project_id:
        description: ProjectID is the ID of the project this sprint belongs to.
        example: 1
        type: integer
      project_name:
        description: ProjectName is the optional name of the associated project.
        example: Website Redesign
        type: string
      start_date:
        description: StartDate is the date when the sprint started.
        example: "2025-04-15T00:00:00Z"
        type: string
      task_count:
        description: TaskCount is the total number of tasks in the sprint (optional).
        example: 3
        type: integer
      tasks:
        description: Tasks is the list of tasks within the sprint (optional).
        items:
          $ref: '#/definitions/dto.TaskInSprintResponse'
        type: array
    type: object
  dto.SprintSliceSuccessResponse:
    properties:
      count:
        example: 5
        type: integer
      data:
        items:
          $ref: '#/definitions/dto.SprintResponse'
        type: array
      message:
        example: Items found successfully
        type: string
    type: object
  dto.SprintSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.SprintResponse'
      message:
        example: Operation successful
        type: string
    type: object
  dto.TaskInSprintResponse:
    properties:
      due_date:
        description: DueDate is the optional due date of the task.
        example: "2025-04-25T00:00:00Z"
        type: string
      id:
        description: ID is the unique identifier of the task.
        example: 101
        type: integer
      priority:
        allOf:
        - $ref: '#/definitions/models.TaskPriority'
        description: Priority is the priority level of the task.
        example: HIGH
      status:
        allOf:
        - $ref: '#/definitions/models.TaskStatus'
        description: Status is the current status of the task.
        example: IN_PROGRESS
      task:
        description: Title is the title or description of the task.
        example: Design homepage layout
        type: string
    type: object
...
  dto.UpdateSprintRequest:
    properties:
      end_date:
        description: EndDate is the optional new end date of the sprint.
        example: "2025-05-05T00:00:00Z"
        type: string
      goal:
        description: Goal is the optional new goal of the sprint.
        example: Finalize UI and start backend integration
        minLength: 5
        type: string
      name:
        description: Name is the optional new name of the sprint.
        example: Sprint 1 - Revised
        minLength: 2
        type: string
      start_date:
        description: StartDate is the optional new start date of the sprint.
        example: "2025-04-20T00:00:00Z"
        type: string
    type: object
...
  models.TaskPriority:
    enum:
    - HIGH
    - MEDIUM
    - LOW
    - CRITICAL
    type: string
    x-enum-varnames:
    - HighPriority
    - MediumPriority
    - LowPriority
    - CriticalPriority
  models.TaskStatus:
    enum:
    - TO_DO
    - IN_PROGRESS
    - REVIEW
    - DONE
    - BLOCKED
    type: string
    x-enum-varnames:
    - ToDoTask
    - InProgressTask
    - ReviewTask
    - DoneTask
    - BlockedTask
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /sprints:
    get:
      description: Retrieves sprints based on optional query parameters (id, name,
        projectid, startdate, enddate)
      parameters:
      - description: Sprint ID
        in: query
        name: id
        type: integer
      - description: Sprint name
        in: query
        name: name
        type: string
      - description: Project ID
        in: query
        name: projectid
        type: integer
      - description: 'Start date after (format: YYYY-MM-DD)'
        in: query
        name: startdate
        type: string
      - description: 'End date before (format: YYYY-MM-DD)'
        in: query
        name: enddate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sprints found
          schema:
            $ref: '#/definitions/dto.SprintSliceSuccessResponse'
        "400":
          description: Bad request - Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Find sprints with filters
      tags:
      - Sprints
    post:
      consumes:
      - application/json
      description: Creates a new sprint for a specific project
      parameters:
      - description: Sprint creation request
        in: body
        name: sprint
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSprintRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Sprint created successfully
          schema:
            $ref: '#/definitions/dto.SprintSuccessResponse'
        "400":
          description: Bad request - Invalid input or dates
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - User not authorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found - Project not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new sprint
      tags:
      - Sprints
  /sprints/{sprintId}:
    delete:
      description: Deletes a specific sprint
      parameters:
      - description: Sprint ID
        in: path
        name: sprintId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sprint deleted successfully
          schema:
            $ref: '#/definitions/dto.GenericSuccessResponse'
        "400":
          description: Bad request - Invalid sprint ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - User not authorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found - Sprint not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a sprint
      tags:
      - Sprints
    get:
      description: Retrieves details of a specific sprint
      parameters:
      - description: Sprint ID
        in: path
        name: sprintId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sprint found
          schema:
            $ref: '#/definitions/dto.SprintSuccessResponse'
        "400":
          description: Bad request - Invalid sprint ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - User not authorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found - Sprint not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a sprint by ID
      tags:
      - Sprints
    put:
      consumes:
      - application/json
      description: Updates the details of an existing sprint
      parameters:
      - description: Sprint ID
        in: path
        name: sprintId
        required: true
        type: integer
      - description: Sprint update request
        in: body
        name: sprint
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSprintRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sprint updated
          schema:
            $ref: '#/definitions/dto.SprintSuccessResponse'
        "400":
          description: Bad request - Invalid input or sprint ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - User not authorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a sprint
      tags:
      - Sprints
...
schemes:
- http
- https
swagger: "2.0"